move pattern functions to plugin file
